Move getSettingsComponent into helpers
Also, add a unit
test for it.
