+++ js/src/edit-block/components/settings/index.js
@@ -1,6 +1,7 @@
 export { default as Checkbox } from './checkbox';
 export { default as Location } from './location';
 export { default as NewLineFormat } from './new-line-format';
+export { default as Number } from './number';
 export { default as NumberNonNegative } from './number-non-negative';
 export { default as Text } from './text';
 export { default as Textarea } from './textarea';
