+++ js/src/edit-block/helpers/index.js
@@ -1,1 +1,2 @@
+export { default as getSettingsComponent } from './getSettingsComponent';
 export { default as initializeEditor } from './initializeEditor';
